+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Support v3 json docs
 * Some overlay corrections to `BlueprintControlBehavior`
+* Automatically manage data and docs links in `Lua.workspace.library`
 
 ## 1.1.20